feedback_agent_flow/pyproject.toml
[project]
name = "feedback_agent_flow"
version = "0.1.0"
description = "feedback_agent_flow using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    "chainlit>=2.4.1",
    "crewai[tools]>=0.95.0,<1.0.0",
    "google-genai>=1.2.0",
    "google-generativeai>=0.8.4",
    "pydantic[email]>=2.10.6",
    "types-pyyaml>=6.0.12.20250326",
]

[project.scripts]
kickoff = "feedback_agent_flow.main:kickoff"
kickoff1 = "feedback_agent_flow.main1:kickoff"
plot = "feedback_agent_flow.main:plot"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "flow"
